update UI and new AI protocol

+++ dnapi/controlitem.h
@@ -11,7 +11,7 @@ class ControlItem : public QObject
     Q_OBJECT
 public:
     Q_PROPERTY(QString name READ name CONSTANT)
-    Q_PROPERTY(int boatID READ boatID CONSTANT)
+    Q_PROPERTY(uint8_t boatID READ boatID CONSTANT)
     explicit ControlItem(QObject *parent, QString name, int controlType, QVector<DNValue> fields);
     ControlItem(const ControlItem& other, QObject *parent = nullptr);
     const ControlItem& operator = (const ControlItem& other);
@@ -22,19 +22,19 @@ public:
     void sendCMD();
     QString name() { return _name; }
     int controlType() { return _controlType;}
-    int boatID() {  return _boatID;}
-    void setBoatID(int boatID) { _boatID = boatID;}
+    uint8_t boatID() {  return _boatID;}
+    void setBoatID(uint8_t boatID) { _boatID = boatID;}
     void setControlType(int type) {_controlType = type;}
     virtual void processMsg(QByteArray command){}
 
 
 signals:
-    void sendMsgbyID(int boatID, uint8_t topic, QByteArray command);
+    void sendMsgbyID(uint8_t boatID, uint8_t topic, QByteArray command);
 private:
     QString _name;
 
     QVector<DNValue*> _fields;
-    int _boatID;
+    uint8_t _boatID;
     int _controlType;
 
 };
